// apps/openldr-web/src/types/database.tsx
import { ColumnDef } from "@tanstack/react-table";

export type Column = {
  id?: string;
  name: string;
  type: string;
  constraint: number | null;
  nullable?: boolean;
  position?: number;
  primaryKey: boolean;
  autogenerated?: boolean;
  optional?: any;
};

export type Table = {
  id: string;
  name: string;
  columns: Column[];
};

export type Database = {
  id: string;
  name: string;
  tables: Table[];
};

export type DataOptions = {
  page: number;
  pages: number;
  returned: number;
  limit: number;
};

export type FilterOption = {
  id: string;
  column: string;
  operator: string;
  combineWith: string;
  value: any | undefined | null;
};

export type SortingOption = {
  id: string;
  column: string;
  ascending: boolean;
};

export type ProjectionOption = {
  id: string;
  column: string;
};

export type CustomColumnDef<TData, TValue> = ColumnDef<TData, TValue> & {
  className?: string;
  customComponent?: React.ReactElement;
};